Prepare 0.5.7

+++ src/core/atomicdex/utilities/kill.cpp
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright © 2013-2021 The Komodo Platform Developers.                      *
+ * Copyright © 2013-2023 The Komodo Platform Developers.                      *
  *                                                                            *
  * See the AUTHORS, DEVELOPER-AGREEMENT and LICENSE files at                  *
  * the top-level directory of this distribution for the individual copyright  *
@@ -14,6 +14,9 @@
  *                                                                            *
  ******************************************************************************/
 
+#include <fstream>
+#include <iostream>
+
 //! PCH Headers
 #include "atomicdex/pch.hpp"
 
@@ -26,11 +29,30 @@ namespace atomic_dex
     kill_executable(const char* exec_name)
     {
 #if defined(__APPLE__) || defined(__linux__)
-        std::string cmd_line = "killall " + std::string(exec_name);
-        std::system(cmd_line.c_str());
+        std::string cmd_line_check = "pgrep " + std::string(exec_name);
+        std::string response = execute(cmd_line_check);
+        if (response != "")
+        {
+            std::string cmd_line = "killall " + std::string(exec_name);
+            std::string response = execute(cmd_line);
+        }
 #else
         std::string cmd_line = "taskkill /F /IM " + std::string(exec_name) + ".exe /T";
-        std::system(cmd_line.c_str());
+        std::string response = execute(cmd_line);
 #endif
+    }
+
+    std::string
+    execute(const std::string& command)
+    {
+        system((command + " > temp.txt").c_str());
+
+        std::ifstream ifs("temp.txt");
+        std::string ret{ std::istreambuf_iterator<char>(ifs), std::istreambuf_iterator<char>() };
+        ifs.close(); // must close the inout stream so the file can be cleaned up
+        if (std::remove("temp.txt") != 0) {
+            SPDLOG_DEBUG("Error deleting temporary file");
+        }
+        return ret;
     }
 } // namespace atomic_dex
